refactor for multi-language support; add ruby
