+++ cli/src/providers/k8s_hyper_alpine.rs
@@ -2,11 +2,11 @@ use std::rc::Rc;
 use std::sync::Arc;
 
 use clap::crate_version;
-use handlebars::{to_json, Handlebars};
+use handlebars::{Handlebars, to_json};
 use serde::Serialize;
 
 use crate::providers::{Options, Provider, ProviderArtifact, ProviderError};
-use crate::transpiler::{asml, Artifact};
+use crate::transpiler::{Artifact, asml};
 
 pub struct ServiceProvider {
     options: Arc<Options>,
@@ -147,6 +147,7 @@ impl Provider for FunctionProvider {
         match ctx.functions.iter().find(|&f| *f.name == name.clone()) {
             Some(function) => {
                 let service = function.service_name.clone();
+                let language = function.language.clone();
                 let registry_type = self
                     .options
                     .get("registry_type")
@@ -173,8 +174,12 @@ impl Provider for FunctionProvider {
                 let data = FunctionData {
                     base_image_version: crate_version!().to_string(),
                     function_name: function.name.clone(),
-                    handler_name: function.handler_name.clone(),
                     service_name: service.clone(),
+                    handler_name: match function.language.as_str() {
+                        "rust" => format!("{}.wasm.bin", function.name.clone()),
+                        "ruby" => "ruby.wasmu".into(),
+                        _ => "handler".into(),
+                    },
                     iomods: iomods.clone(),
                     has_iomods: iomods.len() > 0,
                     container_registry: ContainerRegistryData {
@@ -206,6 +211,7 @@ impl Provider for FunctionProvider {
                             .unwrap_or(&"us-east-1".to_string())
                             .clone(),
                     },
+                    is_ruby: function.language == "ruby".to_string(),
                 };
                 let data = to_json(data);
 
@@ -257,6 +263,8 @@ pub struct FunctionData {
     pub has_iomods: bool,
     pub container_registry: ContainerRegistryData,
     pub iomods: Vec<IomodContainer>,
+
+    pub is_ruby: bool,
 }
 
 #[derive(Serialize)]
@@ -275,7 +283,10 @@ pub struct IomodContainer {
 }
 
 static DOCKERFILE_TEMPLATE: &str = r#"FROM public.ecr.aws/akkoro/assemblylift/hyper-alpine:{{base_image_version}}
-ADD ./{{function_name}}/{{function_name}}.wasm.bin /opt/assemblylift/handler.wasm.bin
+ENV ASML_WASM_MODULE_NAME {{handler_name}}
+ADD ./{{function_name}}/{{handler_name}} /opt/assemblylift/{{handler_name}}
+{{#if is_ruby}}COPY ./{{function_name}}/ruby-wasm32-wasi /usr/bin
+COPY ./{{function_name}}/rubysrc/* /usr/bin/ruby-wasm32-wasi/src/{{/if}}
 "#;
 
 static SERVICE_TEMPLATE: &str = r#"locals {
@@ -337,7 +348,7 @@ resource random_id {{service_name}}_{{function_name}}_image {
     byte_length = 8
     keepers = {
         dockerfile_hash = filebase64sha256("${path.module}/services/{{service_name}}/{{function_name}}/Dockerfile")
-        wasm_hash       = filebase64sha256("${path.module}/services/{{service_name}}/{{function_name}}/{{function_name}}.wasm.bin")
+        wasm_hash       = filebase64sha256("${path.module}/services/{{service_name}}/{{function_name}}/{{handler_name}}")
         iomods_hash     = data.archive_file.{{service_name}}_{{function_name}}_iomods.output_sha
     }
 }
