+++ core/src/wasm.rs
@@ -60,48 +60,52 @@ where
     R: RuntimeAbi<S> + 'static,
     S: Clone + Send + Sized + 'static,
 {
-    let env = ThreaderEnv::new(registry_tx, status_sender);
+    let threader_env = ThreaderEnv::new(registry_tx, status_sender);
+    let function_env = std::env::var("ASML_FUNCTION_ENV").unwrap_or("default".into());
+    let mut wasi_env = match function_env.as_str() {
+        "ruby" => WasiState::new(module_name.clone())
+            .arg("/src/handler.rb")
+            .env("RUBY_PLATFORM", "wasm32-wasi")
+            .map_dir("/src", "./ruby-wasm32-wasi/src")
+            .expect("could not preopen `src` directory")
+            .map_dir("/usr", "./ruby-wasm32-wasi/usr")
+            .expect("could not map ruby fs")
+            .finalize()
+            .expect("could not init WASI env"),
+        _ => WasiState::new(module_name.clone())
+            .finalize()
+            .expect("could not init WASI env"),
+    };
 
-    let mut wasi_env = WasiState::new(module_name.clone())
-        .arg("/src/handler.rb")
-        .env("RUBY_PLATFORM", "wasm32-wasi")
-        // .preopen_dir("/src")
-        .map_dir("/src", "./ruby-wasm32-wasi/src")
-        .expect("could not preopen `src` directory")
-        .map_dir("/usr", "./ruby-wasm32-wasi/usr")
-        .expect("could not map ruby fs")
-        .finalize()
-        .expect("could not init WASI env");
     let wasi_imports = wasi_env
         .import_object(&module)
         .expect("could not get WASI import object");
-
     let asml_imports = imports! {
         "env" => {
-            "__asml_abi_runtime_log" => Function::new_native_with_env(&store, env.clone(), R::log),
-            "__asml_abi_runtime_success" => Function::new_native_with_env(&store, env.clone(), R::success),
+            "__asml_abi_runtime_log" => Function::new_native_with_env(&store, threader_env.clone(), R::log),
+            "__asml_abi_runtime_success" => Function::new_native_with_env(&store, threader_env.clone(), R::success),
 
-            "__asml_abi_invoke" => Function::new_native_with_env(&store, env.clone(), asml_abi_io_invoke), // TODO deprecated, IOmod guests need to update
-            "__asml_abi_io_invoke" => Function::new_native_with_env(&store, env.clone(), asml_abi_io_invoke),
-            "__asml_abi_io_poll" => Function::new_native_with_env(&store, env.clone(), asml_abi_io_poll),
-            "__asml_abi_io_len" => Function::new_native_with_env(&store, env.clone(), asml_abi_io_len),
-            "__asml_abi_io_load" => Function::new_native_with_env(&store, env.clone(), asml_abi_io_load),
-            "__asml_abi_io_next" => Function::new_native_with_env(&store, env.clone(), asml_abi_io_next),
+            "__asml_abi_invoke" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_io_invoke), // TODO deprecated, IOmod guests need to update
+            "__asml_abi_io_invoke" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_io_invoke),
+            "__asml_abi_io_poll" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_io_poll),
+            "__asml_abi_io_len" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_io_len),
+            "__asml_abi_io_load" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_io_load),
+            "__asml_abi_io_next" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_io_next),
 
-            "__asml_abi_clock_time_get" => Function::new_native_with_env(&store, env.clone(), asml_abi_clock_time_get),
+            "__asml_abi_clock_time_get" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_clock_time_get),
 
-            "__asml_abi_input_start" => Function::new_native_with_env(&store, env.clone(), asml_abi_input_start),
-            "__asml_abi_input_next" => Function::new_native_with_env(&store, env.clone(), asml_abi_input_next),
-            "__asml_abi_input_length_get" => Function::new_native_with_env(&store, env.clone(), asml_abi_input_length_get),
+            "__asml_abi_input_start" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_input_start),
+            "__asml_abi_input_next" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_input_next),
+            "__asml_abi_input_length_get" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_input_length_get),
 
-            "__asml_expabi_z85_encode" => Function::new_native_with_env(&store, env.clone(), asml_abi_z85_encode),
-            "__asml_expabi_z85_decode" => Function::new_native_with_env(&store, env.clone(), asml_abi_z85_decode),
+            "__asml_expabi_z85_encode" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_z85_encode),
+            "__asml_expabi_z85_decode" => Function::new_native_with_env(&store, threader_env.clone(), asml_abi_z85_decode),
         },
     };
 
     let import_object: Resolver = asml_imports.chain_back(wasi_imports);
 
-    Ok((import_object, env))
+    Ok((import_object, threader_env))
 }
 
 pub fn new_instance(
